refactor(ui): move RepostersOverlay into GroupedRepostHeader subfolder
Made-with: Cursor
